+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * [#8508](https://github.com/rubocop-hq/rubocop/pull/8508): Fix a false positive for `Style/CaseLikeIf` when conditional contains comparison with a class. Mark `Style/CaseLikeIf` as not safe. ([@fatkodima][])
 * [#8534](https://github.com/rubocop-hq/rubocop/issues/8534): Fix `Lint/BinaryOperatorWithIdenticalOperands` for binary operators used as unary operators. ([@marcandre][])
 * [#8537](https://github.com/rubocop-hq/rubocop/pull/8537): Allow a trailing comment as a description comment for `Bundler/GemComment`. ([@pocke][])
+* [#8533](https://github.com/rubocop-hq/rubocop/pull/8533): Fix a false positive for `Lint/BinaryOperatorWithIdenticalOperands` when using `*` operator with same variable operand. ([@koic][])
 
 ### Changes